+++ src/AvroConvert/AvroConvert.Json2Avro.cs
@@ -21,13 +21,38 @@
 #endregion
 
 using Newtonsoft.Json;
+using SolTechnology.Avro.Features.JsonToAvro;
 
 namespace SolTechnology.Avro
 {
     public static partial class AvroConvert
     {
         /// <summary>
-        /// Converts Json object directly to Avro format
+        /// Converts JSON object directly to Avro format
+        /// Warning! This is an experimental feature.
+        /// </summary>
+        public static byte[] Json2Avro(string json)
+        {
+            var decoder = new JsonToAvroDecoder();
+            return decoder.DecodeJson(json, CodecType.Null);
+
+        }
+
+
+        /// <summary>
+        ///  Converts JSON object directly to Avro format
+        /// Choosing <paramref name="codecType"/> reduces output object size
+        /// Warning! This is an experimental feature.
+        /// </summary>
+        public static byte[] Json2Avro(string json, CodecType codecType)
+        {
+            var decoder = new JsonToAvroDecoder();
+            return decoder.DecodeJson(json, codecType);
+        }
+
+
+        /// <summary>
+        ///  Converts JSON object directly to Avro format
         /// </summary>
         public static byte[] Json2Avro<T>(string json)
         {
@@ -39,8 +64,8 @@ namespace SolTechnology.Avro
 
 
         /// <summary>
-        ///  Converts Json object directly to Avro format
-        /// Choosing <paramref name="codecType"/> reduces output object size
+        ///  Converts JSON object directly to Avro format
+        ///  Choosing <paramref name="codecType"/> reduces output object size
         /// </summary>
         public static byte[] Json2Avro<T>(string json, CodecType codecType)
         {