+++ src/AvroConvert/AvroConvert.SerializeHeadless.cs
@@ -26,7 +26,7 @@ namespace SolTechnology.Avro
     public static partial class AvroConvert
     {
         /// <summary>
-        /// Serializes given object to AVRO format - <c>excluding</c> header
+        /// Serializes given object to Avro format - <c>excluding</c> header
         /// </summary>
         public static byte[] SerializeHeadless(object obj, string schema)
         {